Add WoodType package and create a utility class WoodTyper
The WoodType package contains classes that extend WoodItem and initialize
themselves with the appropriate constants (type, base delivery time, and
price).

The WoodTyper is a static utility class that instatiates a WoodItem based on
the type of wood. This allows us to easily use one class (WoodItem) to handle
all of our wood types, but allow them to easily be initialized with the
required constants.