Add additional classes for storing a customer order
Created several classes to handle an order to be created from an input
file. In addition, the function to read the input file and convert the
data into the Order object has been implemented. Finally, an example input
file has been provided.

In the future, the input file should be available to be input through a
command line argument.